chore: s3_utils_layer integration
