+++ lambda/api_fetch_videos_handler/pyproject.toml
@@ -9,5 +9,6 @@ dependencies = []
 extraPaths = [
   "../layers/vector_database_layer/",
   "../layers/response_utils_layer/",
+  "../layers/s3_utils_layer/",
   "../layers/config_layer/",
 ]
