+++ lambda/sqs_embedding_task_producer/lambda_function.py
@@ -1,14 +1,12 @@
 import json
 import boto3
-import time
-import botocore.exceptions
 import os
 import logging
 import utils
-import urllib.parse
 from embed_service import EmbedService
 from config import get_secret, setup_logging, get_db_config
 from vector_db_service import VectorDBService
+import s3_utils
 
 # Environment variables
 SECRET_NAME = os.getenv("SECRET_NAME", "kubrick_secret")
@@ -21,57 +19,12 @@ VIDEO_EMBEDDING_SCOPES = json.loads(
     os.getenv("VIDEO_EMBEDDING_SCOPES", '["clip", "video"]')
 )
 QUEUE_URL = os.environ["QUEUE_URL"]
-S3_REGION = os.getenv("S3_REGION", "us-east-2")
+S3_REGION = os.getenv("S3_REGION", "us-east-1")
 
-s3 = boto3.client("s3", region_name=S3_REGION)
 sqs = boto3.client("sqs")
 logger = setup_logging()
 SECRET = get_secret(SECRET_NAME)
 DB_CONFIG = get_db_config(SECRET)
-
-
-def wait_for_file(
-    s3_client,
-    bucket,
-    key,
-    retries=FILE_CHECK_RETRIES,
-    delay=FILE_CHECK_DELAY_SEC,
-    logger=logging.getLogger(),
-):
-    for attempt in range(retries):
-        try:
-            logger.info(
-                f"Checking if file exists: s3://{bucket}/{key} (attempt {attempt + 1}/{retries})"
-            )
-            s3_client.head_object(Bucket=bucket, Key=key)
-            logger.info("File found in S3")
-            return True
-        except botocore.exceptions.ClientError as e:
-            if e.response.get("Error", {}).get("Code") == "404":
-                logger.warning(f"File not found. Retrying in {delay} seconds...")
-                time.sleep(delay)
-            else:
-                logger.error(f"Unexpected error while checking file existence: {e}")
-                raise
-    logger.warning("File still not found after all retries")
-    return False
-
-
-def extract_s3_info(event):
-    records = event.get("Records", [])
-    if not records:
-        raise ValueError("No Records found in event")
-
-    record = records[0]
-    bucket = record.get("s3", {}).get("bucket", {}).get("name")
-    key = urllib.parse.unquote_plus(
-        record.get("s3", {}).get("object", {}).get("key", "")
-    ).strip()
-
-    if not bucket or not key:
-        raise ValueError("Missing bucket or key in event")
-
-    return bucket, key
 
 
 def persist_task_metadata(
@@ -99,7 +52,7 @@ def lambda_handler(event, context):
     vector_db_service = VectorDBService(db_params=DB_CONFIG, logger=logger)
 
     try:
-        bucket, key = extract_s3_info(event)
+        bucket, key = s3_utils.extract_s3_info(event)
         logger.info(f"Extracted S3 bucket: {bucket}, key: {key}")
 
         if key.endswith("/"):
@@ -110,8 +63,7 @@ def lambda_handler(event, context):
             raise ValueError("File is not a video or the video format is not supported")
         logger.info("Validated video file type")
 
-        if not wait_for_file(
-            s3,
+        if not s3_utils.wait_for_file(
             bucket,
             key,
             FILE_CHECK_RETRIES,
@@ -120,11 +72,10 @@ def lambda_handler(event, context):
         ):
             raise FileNotFoundError(f"File s3://{bucket}/{key} not found after retries")
 
-        presigned_url = s3.generate_presigned_url(
-            "get_object",
-            Params={"Bucket": bucket, "Key": key},
-            ExpiresIn=PRESIGNED_URL_TTL,
+        presigned_url = s3_utils.generate_presigned_url(
+            bucket=bucket, key=key, expires_in=PRESIGNED_URL_TTL
         )
+
         logger.info(f"Presigned URL generated (expires in {PRESIGNED_URL_TTL} seconds)")
         logger.debug(f"presigned_url={presigned_url}")
