+++ lambda/sqs_embedding_task_producer/pyproject.toml
@@ -4,13 +4,12 @@ version = "0.1.0"
 description = "SQS Embedding Task Producer Lambda function"
 readme = ""
 requires-python = ">=3.13"
-dependencies = [
-    "boto3>=1.39.17",
-]
+dependencies = ["boto3>=1.39.17"]
 
 [tool.pyright]
 extraPaths = [
   "../layers/vector_database_layer/",
   "../layers/embed_service_layer/",
   "../layers/config_layer/",
+  "../layers/s3_utils_layer/",
 ]
